+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Added ToCollection extension method for copying an enumerable to a new collection
+
 ## [11.0.0] - 2020-03-20
 ### Added
  - Added Not null tag for ICDUriBuilder Constructor that takes a URI as an argument.